requirements.txt
python-ffmpeg==2.0.4
argparse==1.4.0 
requests>=2.31.0
pathlib>=1.0.1
PyYAML>=6.0.1
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6 